Move remediation functionality to pkg/remediation
